move cart to /pages
