+++ caterly-client-frontend/src/App.tsx
@@ -3,7 +3,7 @@ import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import Layout from "./components/layout/Layout";
 import LoginForm from "./pages/login/LoginForm";
 import FoodListPage from "./pages/FoodListPage";
-import Cart from "./components/cart/Cart";
+import Cart from "./pages/Cart";
 import { QueryClientProvider } from "@tanstack/react-query";
 import { queryClient } from "./lib/queryClient";
 import RegistrationForm from "./pages/registration/RegistrationForm";
